backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

sqlalchemy==2.0.34
asyncpg==0.29.0
alembic==1.13.2

pydantic==2.9.2
pydantic-settings==2.5.2

httpx==0.27.2
feedparser==6.0.11
python-dateutil==2.9.0.post0
